!htp Move `ClassMagnet` from `server.directives` into `server.util` package
